+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Supply Chain DApp - ChangeLog
 
+## [0.2.0] - 18/02/2025
+
+- Add dependencies :
+   - @nomicfoundation/hardhat-ethers
+- Update hardat.config
+- Create Hardhat script to deployed contract
+- Create React front-end to connect at the hardat server
+
 ## [0.1.0] - 06/02/2025
 
  - Add dependencies :
    - ethers ^6.13.5
    - hardhat ^2.22.18
    - web3 ^4.16.0
- - Add contract solidity
- - Add Hardhat test
+ - Create contract solidity
+ - Create Hardhat test